requirements.txt
python-dotenv==1.0.1
requests==2.32.3
feedparser==6.0.11
trafilatura==1.12.2
beautifulsoup4==4.12.3
tqdm==4.66.4
jinja2==3.1.4
sendgrid==6.11.0
openai==1.40.0
